Force utc_offset to use local time.
This is a shim to deal with ActiveSupport::TimeWithZone objects in
Rails. to_s_for_cap is not defined for these objects, so ActiveSupport
delegates the call to a new Time object. In the conversion from
TimeWithZone to Time, the timezone offset is lost. Using localtime gets
it back again.